update image with manager
