+++ operator/bundle/manifests/multicluster-global-hub-operator.clusterserviceversion.yaml
@@ -636,6 +636,8 @@ spec:
                   value: quay.io/stolostron/multicluster-global-hub-manager:latest
                 - name: OPERAND_IMAGE_MULTICLUSTER_GLOBAL_HUB_AGENT
                   value: quay.io/stolostron/multicluster-global-hub-agent:latest
+                - name: OPERAND_IMAGE_GRAFANA
+                  value: quay.io/stolostron/grafana:2.8.0-SNAPSHOT-2023-03-06-01-52-34
                 - name: OPERAND_IMAGE_OAUTH_PROXY
                   value: quay.io/stolostron/origin-oauth-proxy:4.9
                 image: quay.io/stolostron/multicluster-global-hub-operator:latest
