+++ operator/pkg/controllers/hubofhubs/system_config.go
@@ -1,0 +1,100 @@
+package hubofhubs
+
+import (
+	"context"
+	"strconv"
+
+	operatorv1alpha2 "github.com/stolostron/multicluster-global-hub/operator/apis/v1alpha2"
+	"github.com/stolostron/multicluster-global-hub/operator/pkg/config"
+	"github.com/stolostron/multicluster-global-hub/operator/pkg/utils"
+	"github.com/stolostron/multicluster-global-hub/pkg/constants"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/equality"
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+)
+
+// reconcileSystemConfig tries to create hoh resources if they don't exist
+func (r *MulticlusterGlobalHubReconciler) reconcileSystemConfig(ctx context.Context,
+	mgh *operatorv1alpha2.MulticlusterGlobalHub,
+) error {
+	// check for image overrides configmap
+	var imageOverridesConfigmap *corev1.ConfigMap
+	if imageOverridesConfigmapName := config.GetImageOverridesConfigmap(mgh); imageOverridesConfigmapName != "" {
+		var err error
+		imageOverridesConfigmap, err = r.KubeClient.CoreV1().ConfigMaps(mgh.GetNamespace()).Get(
+			ctx, imageOverridesConfigmapName, metav1.GetOptions{})
+		if err != nil {
+			return err
+		}
+	}
+
+	// set imgae overrides
+	if err := config.SetImageOverrides(mgh, imageOverridesConfigmap); err != nil {
+		return err
+	}
+
+	if err := r.Client.Get(ctx,
+		types.NamespacedName{
+			Name: constants.GHSystemNamespace,
+		}, &corev1.Namespace{}); err != nil {
+		if errors.IsNotFound(err) {
+			if err := r.Client.Create(ctx, &corev1.Namespace{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: constants.GHSystemNamespace,
+					Labels: map[string]string{
+						constants.GlobalHubOwnerLabelKey: constants.GHOperatorOwnerLabelVal,
+					},
+				},
+			}); err != nil {
+				return err
+			}
+		} else {
+			return err
+		}
+	}
+
+	// hoh configmap
+	hohConfigMap := &corev1.ConfigMap{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: constants.GHSystemNamespace,
+			Name:      constants.GHConfigCMName,
+			Labels: map[string]string{
+				constants.GlobalHubOwnerLabelKey:       constants.GHOperatorOwnerLabelVal,
+				constants.GlobalHubGlobalResourceLabel: "",
+			},
+		},
+		Data: map[string]string{
+			"aggregationLevel":    string(mgh.Spec.AggregationLevel),
+			"enableLocalPolicies": strconv.FormatBool(mgh.Spec.EnableLocalPolicies),
+		},
+	}
+
+	existingHoHConfigMap := &corev1.ConfigMap{}
+	if err := r.Client.Get(ctx,
+		types.NamespacedName{
+			Namespace: constants.GHSystemNamespace,
+			Name:      constants.GHConfigCMName,
+		}, existingHoHConfigMap); err != nil {
+		if errors.IsNotFound(err) {
+			if err := r.Client.Create(ctx, hohConfigMap); err != nil {
+				return err
+			}
+			return nil
+		} else {
+			return err
+		}
+	}
+
+	if !equality.Semantic.DeepDerivative(hohConfigMap.Data, existingHoHConfigMap.Data) ||
+		!equality.Semantic.DeepDerivative(hohConfigMap.GetLabels(), existingHoHConfigMap.GetLabels()) {
+		hohConfigMap.ObjectMeta.ResourceVersion = existingHoHConfigMap.ObjectMeta.ResourceVersion
+		if err := utils.UpdateObject(ctx, r.Client, hohConfigMap); err != nil {
+			return err
+		}
+		return nil
+	}
+
+	return nil
+}
